feat: #10 CookFlow 거지 진척도 반영 외 완성 수정, 추가해야할 점 - 카메라 가로 불가 버그 - 레시피 추천 시 장바구니에 담은 재료 우선 추천 - 레시피 추천 함수가 앱을 킬 때만 돌아가는데, 재료를 추가, 삭제할 때도 가능하게 해야 함.

+++ BeggarsGrowing/BeggarsGrowing/DataModel/RefrigeratorModel.swift
@@ -19,11 +19,11 @@ import SwiftData
     /// 냉동 여부
     var freezing: Bool
     /// 현재 잔량
-    var amount: Double
+    var amount: Int
     /// 등록 날짜
     var date: Date
     
-    init(food: String, price: Int, amount: Double, freezing: Bool, date: Date) {
+    init(food: String, price: Int, amount: Int, freezing: Bool, date: Date) {
         self.food = food
         self.price = price
         self.amount = amount
